Choke on invalid command options.
Even if there are no options supported, if one is passed, complain. Add tests to ensure we complain, to show the help output for the proper command (this was broken), and that the args param is cleared of options after they are parsed out. We will pass any remaining arguments to `execute()`.