Use IPC::System::Simple instead of our own system/capture code.
This is much cleaner, except for the bit where I clean out the " at line ..." part of the message. Still, much less to worry about, portability-wise. And as a benefit, we now have proper exception handling for failures of both `run()` and `capture()`.

Oh, and `do_system()` was renamed `run` and `backticks()` qwas renamed `capture()`. Nicer terms, I think.

Came about as I wrestled with the pg engine, so there are some changes and additional tests there, but it's very much a work in progress.