Remove hardcoded #!/usr/bin/perl from bin/sqitch
A hardcoded #!/usr/bin/perl in any script can be problematic with
cpantesters, as it can force use of the system perl instead of the
testing perl. It can also be an issue when installed on systems that
don't have a perl in /usr/bin.

The alternative is a plain #!perl which Module::Build converts at Build
time to an appropriate path. Since bin/sqitch is nothing more than a
shell around App::Sqitch we create an alternative t/sqitch with
"#!/usr/bin/env perl", which works as expected when switching Perls
with perlbrew.