Add methods to begin and rollback tag deploys and reverts.
Also add methods to check for `requires` and `conflicts` issues and cancel the
deploy if there are issues.

Also, abandon using IPC to communicate with `psql` in favor of using DBD::Pg.
This is vastly simpler and less error-prone. Plus errors are real and
immeidate, and there are no quotign issues to worry about. This after spending
serveral days wrestling with IPC, posting about it [StackOverflow][] and
[PerlMonks][], and finally [a blog post][]. Thanks to @mlawren for the dose of
sanity there.

Lots more testing of the pg engine to come, but at least there is now a solid
plan in place, so development should speed up again.

[StackOverflow]: http://stackoverflow.com/q/10569805/79202
[PerlMonks]: http://www.perlmonks.org/?node_id=970244
[a blog post]: http://justatheory.com/computers/databases/dbi-in-sqitch.html