Change schema from step history to deploy events.
That is, rather than just track step deploys and reverts in a "history" table,
track all kinds of events in an "events" table. There are five events:
"deploy", "revert", and "fail" for ssteps, and "apply" and "remove" for tags.
Modify the code in pg.pm to log all except step failure, which will come next.