Finish rewriting Plan.
Now tracks all lines to preserver formatting and comments. Tracks lines and
nodes (steps and tags) in two different places, so that the file can be
written with comments, but iteration is only over steps and tags. Switch to
using Array::AsHash for tracking them, so that we can look things up by tag
and step name (hash) and iterate in order (array). Will also make it easier to
add nodes to the plan in the forthcoming `plan` command.

No longer dealing with untracked, as there will be no such thing. Either a
step is known to the plan or it isn't. The plan file will be required and
authoritative.

Iteration now includes both steps and tags. Tags are no longer containers
for steps, but separate obects with a single name.

Not currently sorting steps by dependecies; need to do that next.