Add LineList and simplify NodeList API.
They fundamentally need to be treated differently. So separate them. This
allows the API to be simpler, too: we literally pass a list of nodes or lines,
rather than a list of key/value pairs as was required by Array::AsHash. This
will also allow for more node-specific code in NodeList.