Add `info` to Tag and Step and use for IDs.
The Tag info is made up of the project URI, the tag name, and the ID of the
step the tag applies to. This allows it to be unique for a project, but not to
change unless the tag is moved to another step.

The Step info is made up of the project URI, the step name, and the ID of the
tag preceeding the step, if there is one. This again allows it to be unique
per project and tag, even if the contents of the deploy script change or any
of the dependencies change.

In passing, remove the default null SHA1 for the step ID in tag. Should
probably change things so that a tag can't be applied if there are no steps in
the plan.