Index nodes by ID in NodeList.
This requires changes to a bunch of tests that previously were not setting a
proect URI.

To allow nodes to be looked up by ID without possibility of conflicting with
step names, also disallow names that match `/^[0-9a-f]{40}/`. Unlikely to
happen anyway, but best to be safe.

In passing, fix a bug in ID generation where the string to be hashed was not
first encoded as UTF-8.