Remove `latest_item` and `latest_tag` and add `latest_step_id`.
Then have `_sync_plan()` rely on them. This is much simpler, a single lookup,
and no need to look up tags separately.

This turned up a situation where the "users" step was in the test plan twice.
Not sure how this happened, but I replaced it with a new "lolz" step and all
is well.