Add `is_dupe` to Step and change file nams.
If a step is later duped, its file name should contain `@$tag_name` in it. So
we add an `is_dupe` flag and, if it is true, add the tag name. Done by
refactoring the generation of the file base name into a new `_fn` attribute,
just to keep things simple.