Add `@ROOT` as an implicit tag on the first step.
Requires reserving "ROOT" as a name, of course, just like tag, for which this
is the complement.

In the process, fix lookups of `@HEAD` in StepList so that it *should* always
work properly. Probably should have noticed it before. At any rate, `@HEAD`
should now always refer to the last step in the plan, while `@ROOT` should
always refer to the first, and both should work for lookups.