Include @ROOT ID in info for untagged steps.
Except for the first one, of course. Because it is since said tag. Needed to
add `_root_step` to plan to make it work, and it is set as a side-effect of
`_parse`, which is a little hinky: if one was to call it before accessing
`_plan`, it would return `undef`; but as it is private, no one else should do
that.

I might change it to `_root_tag` to avoid the overhead of creating a new tag
object for the info of every step that needs it. Tomorrow.