Remove @ROOT from all step info and IDs.
Why? Two reasons:

1. It is an implicit tag, not an actual tag.
2. Due to dependency resolution, the order of steps can change, which means
   that if you have step1 and step2, and step1 is changed to depend on step2,
   then step2 will become the first step in the plan, and so the IDs of the
   two steps would change.

So I think it was just a bit too magical, without any real gain. Good to keep
@ROOT around as an implicit tag for lookups and such, though.

Making the change as a new commit, rather than just throwing out the previous
commit, so as to keep a record around in case I change my mind in the future
and want to put it back quickly. Unlikely, but you never know.