Mark previously-seen step as duped.
Was a bug. When two steps had the same name, was marking the second one as
duped. This was wrong: it needs to be the earlier one that's marked as duped.
This allows the earlier one to find its file as of the first tag that appears
between them, while the dupe still refers to the un-quailified file name.