Refactor dependency parsing into its own section.
This eliminates the need for `(?{})` blocks, which have strange side-effects
and must be carefully written to deal with backtracking anyway (and this code
was not, as pointed out by sprout [#19]). Also, the pattern I had was
imperfect, in that it was bending over backwards to capture only the first
blanks before the dependency list and thus making intervening blanks
*optional*, which is undesireable (as pointed out by Darin McBride on the Perl
5 Porters list).

So now we just capture the entire list of dependencies as a single string, and
parse it separately if we need it. This has the added benefit of complaining
about syntax errors for dependencies independent of the entire step line.

Resolves #19.