Add `rework` command.
It handles additional dependencies and properly builds on the
previously-existing step's files by copying the original scripts to new files.
If the original script does not have a particular file, it of course won't be
copied. No errors, either.

In passing, have `plan` handle calls to `sqitch` in lines (changes, tags,
etc.).