Add `current_state` to Engine and implement in pg.
Not very keen on having to require DateTime, but the alternative is to do all
sorts of hacking to convert things to the local time zone -- including how to
figure out what the local time zone is! Better just to let DateTime do all
that, since it's the sort of thing it was designed for and does it very well.

Will be using this in the forthcoming `status` command.