Return iterator code references instead of lists.
In Engine. This is because they could be very long. Of course, `status` just
reads the whole list anyway (so that it can calculate how long to make the
first column), but since the forthcoming log query function will potentially
return a *very* long list of results, it seemed best to settle on a simple
iterator interface now.