Add timestmap and planner to Change and Tag.
The same information appears in the text representation of each, as well,
separated by spaces.

Also, remove the project URL from the info for these objects, as well as the
tag preceding the change, and put in the timestamp and planner info. The IDs
will still end up unique unless the same user adds a change to two projects
with the same name at the same second.

Tests are failing in `t/change.t` because the parser has not yet been updated.