Add timestamp and planner support to the plan parser.
The timestamp and planner information are now both required for everh tag and
change in the plan file. To keep things unambiguous, dependencies are now
required to be wrapped in brackets. The timestamp is strict ISO-8601 format in
UTC (with trailing Z). The planner's name may no contain "<" or start with
"[", to disambiguate it from the email address and the dependencies. And
the email address may not contain ">", so that we find the end of it (the
file wraps it in <>).

This change itself was not a huge issue. It required all the preceding commits
that changed Tag and Change, but otherwise, this commit updates all the test
plans to contain the necessary timestmap and planner information, and updates
tests to recognize that data. The `t/plan.t` test in particular required quite
a bit of refactoring, as the ad-hoc creation of tag and change objects needed
a bit more flexibility (the utility functions that create them now require
hash ref params instead of a list).