Change note requirement to a method that does the work.
It just made no sense to do it in `BUILDARGS`, because `note_prompt()`
actually needs to be able to call accessors to display information, especially
in Change (forthcoming commit). So just move it to a method. The commands will
call the method after ading a change or tag, but before writing the plan file.