Add `:event` arg to `%C` format.
And remove the optional true argument to `%e`, `%l`, and `%L`. This is
cleaner, a better place, and allows the color to be used in the display of
some text other than the event type. For example, we might later decide that
the change ID should be in the selected color, rather than always be yellow.