Integrate App::Sqitch::Plan::Dep.
It is now used in Plan, Change, and pg. May or may not end up turning
dependencies fetched from the database into Depend objects, not sure yet. So
far we are just using them as strings, so no big deal there. Will change it
if the cross-project dependency stuff requires it.