Require plan pragmas in header section.
Now that dependencies must always have a defined project name, we need to make
sure that the pragma appears before the first dependency declaration. Seems
like we ought to just have them all declared up-front. So change the plan
format to require pragmas at the top of the file, with a blank line before
starting the list of changes.

And now that we're guaranteed to have the project name defined, pass it, and
the plan object itself, to the Depend constructor. And since the current
project's is always set, we have to compare project names while sorting in
order to skip dependencies on changes in other projects.

And finally, update the tests to make them again pass with all the
newly-required attributes of the Depend object.