Support change IDs in dependency specifications.
The hash must be the full hash, 40 charaters long and consisting of only valid
hash charaters, otherwise it's considered a change name. IDs passed to the
constructor are mutually exclusive with change names and tags, although they
may be project-qualified. If a change is a local change, C<id> should always
return the ID. The C<key_name> method will return the ID only if there is no
change or tag.