Add `is_external` to Depend.
Now that dependencies can be spcified with IDs, the project name is optional.
So we update the code to reflect that, but try to get a default from the
current plan if the dependency is unambiguously local.

To keep things clean and prevent infinite recursion, the new `_id_passed`
attribute is used to indicate that the ID was passed to the constructor, and
so it's possible to fetch it without falling back on the default, which is
fetched from the plan.