Rename `_id_passed` to `got_id()` and add `got_project()`.
These will be used by the plan during parsing to determine how to look up
prerequisites without recursing back to the plan object itself (because `id`
and `project` do that and depend on one another). Yes, this is a bit hinky,
but was the cleanest way I could think to do it while providing public APIs
and keeping the complexity of the recursive calls between Plan and Depend
during a parse completely contained in the Plan object.