Start updating tutorial with project and schema.
That is, instead of creating a role, which is a global object and therefore
causes some annoying side-effects requring that databases be dropped so the
role can be dropped, instead create a schema and put everything in it. It's
not a global object, so the complexity goes away. Good for keeping the
tutorial sooner.

Anyway, updated examples and output up to the "Flip Out" section. Will
finish that next.