Attempt to fix user name issues.
On some systems, `getlogin` returns nothing. Which is annoying. Furthermore,
Windows Perl issues a warning that `getpwnam` is not implemented. So try a
bunch of different approaches to determing the current system username, and
then use that value (stored in the new `sysuser` attribute) to get the full
name via Win32API::Net on Windows and via User::pwent elsewhere.