Replace `_dir_map` with individual dest attributes.
So now `dest_top_dir`, `dest_deploy_dir`, `dest_revert_dir`, and
`dest_test_dir` are the places where we should get the names of the individual
destination directories.

Also, put them to use in `bundle_plan()` and `bundle_scripts()` thus putting
the files in the actual proper locations (which was broken).