Fix RPM versions.
I did not realize that RPM versions were dotted integer rather than decimal.
So while 0.92 is newer than 0.911 to Perl, it is not to RPM. So add a macro so
the spec files to tell RPM that 0.92 should be considered 0.920.

While at it, just rely on that version number in sqitch-pg to set the
requirement for the sqitch RPM.