Fix module requirement declarations.
I did not know that minimum versions declared on a `use` line should *not* be
followed by a comma before listing stuff to import. Turns out that Exporter
properly determines if the first argument is numeric and uses that to check
the dependency. This is nice of it, but then the AutoPrereqs dzil plugin
failed to notice it. So some dependencies may be incorrectly listed.
Term::ANSIColor was one that showed up with a failed test from cpan-testers.

So remove the commas, and add some versions where they had been omitted, plus
update the version for Config::GitLike in `dist/sqitch.spec`. Oh, and use
`parent` instead of `base` in the Module::Build subclass.