Allow more charcters to be used before digits at the end of names.
I realized that lot of folks might want to use tags such as "v1.3-1" or
"v23+1". So let them. The only punctuation characters no longer allowed to
be used before digits at the end of a name are `/[~^/=%]/`.

To better document the allowed format of name, describe the rules in
`sqitchchanges.pod` and provide examples of good and bad names. Then
link to it from `sqitch-add.pod`.