Start adding l10n infrastructure.
Language files are in the `po` directory. The Makefile there can be used
to update them. New ones can be created by copying `App-Sqitch.po` and
adding the new language code to `PACKAGE`.

We set the locale in the scripts so that Text::LocaleDomain knows where what
language to use when Sqitch runs.