Convert command option key dashes to underscores.
Noticed that --dir was ignored by the `bundle` command. Turns out that
`_parse_opts()` in Command.pm was not converting dashes to underscores in
option keys. The docs said that it would, but in some places commands expected
dashes and in others they expected underscores.

So normalize them all to have underscores, and make the commands always
expect underscores.