Have `bundle` copy scripts in subdirectories.
This is a bug fix. Previously, `bundle` would copy scripts to the top-level of
the deploy, revert, and test directories. Then, when running `sqitch deploy`
in the bundled directory, Sqitch sould not be able to find the scripts! So use
the new `path_segments()` method added in 162473b to generate the bundle
destination paths, rather than the base name of the scripts.

This commit includes tests for this functionality.