More l10n-related fixes.
I failed to notice in some places that `__` does not have a signnature of
`($), so it was sucking up stuff after it, especially in tests, but also in a
couple of places in log.pm. So change those to `__()` to fix that problem. All
tests now pass with `export LANGUAGE=fr`, so cool!