Force Locale::TextDomain to encode in UTF-8 and to decode all messages.
Based on a discussion with Locale::TextDomain maintainer Guido Flohr. He is
adamantly opposed to returning decoded strings, but points out that through
the combination of setting the `$OUTPUT_CHARSET` environment variable and
binding a filter to the domain, we can get it to properly decode strings. So
let's do that.

Reported in the discussion of pull request #46.