Preserve original plan order when dependencies allow it.
The original
[`tsort`](http://alumnus.caltech.edu/~copeland/work/tsort.html)-derived
algorithm did not attempt to preseve the original order of the list of
changes. My thanks to @user5402 for providing the solution to fix this issue
in [this SO answer](http://stackoverflow.com/a/13370233/79202). As a
side-effect, a couple of other tests that had been listing changes in the
order they are specified as dependencies now preserve the original order
there, too.