Improve failed change log error handling.
I was getting a failure that just said the transaction was rolled back, with
no notice of the original error. This was happening because the deploy script
properly ran, but Sqitch's logging of the change failed.

So we do exception handling around the logging of the change. If it dies,
Sqitch now vents the underlying error, tells the engine to roll back work, and
then runs the revert script so as to restore the database to its original
state.