Suppress duplicate "Deploy failed" message.
In the "tag" and "all" modes, after a deployment fails, prior changes are
reverted and then a final "Deploy failed" message is emitted. So we ended up
with two of those emitted when the failure was due to a logging failure. So
fix it by setting the ident for the logging failure exception to "private" and
suppressing it in the "all" and "tag" modes.