Stop reordering changes when parsing the plan.
They were able to be reordered based on dependencies, but this led to
different orders in different runs, especially if the plan was changed since
the last run.

Instead, throw an exception if the order is invalid. But work hard to suggest
to the user how to edit the plan to fix the issue by tracking ordering issues
and providing a hint with the recommendation to move a change x lines in the
plan file to fix the problem.