Return max plan index when updating old IDs.
Because, when we had the old IDs, things could be run in an order different
from the plan, the last run change may actually appear before other deployed
changes in the plan. So, as we updated the IDs, figure out which of the
deployed changes appears last in the plan, and use *that* change as the
current position.

Ref #56.