Calculate relative changes to @LAST and @FIRST in the database.
Previously, the code was finding the last or first change in the database,
then caclucating the offset determined by `~` or `^` relative to the change's
place in the plan. This could be a little surprising if the plan changed and
one was trying to revert to, say, `@LAST^`. So instead, apply the offsets
determined by `~` or `^` at the database level.

Currently, this only works for `@LAST` and `@FIRST`. Other changes are still
looked up relative to the plan. This still might be suprising in the case of,
say, `sqitch revert foo^`. I think that case also requires that the `^` be
calculated in the database. But only for `reverts`, I think.