Change "test" to "verify".
I want it to be conceptually different from the standard idea of a test, which
is mostly about unit testing. I started to use "check", but I had already
planned to use that term for sanity-checking a project plan, scripts,
deployments, etc. I was going to do "accept", because these are effectively
acceptance tests, but `sqitch accept` was wrong, because it wouldn't accept
anything, but run acceptance tests.

The term "verify" works well, though. It nicely complements "deploy" and
"revert" (it even has the same number of letters!), and `sqitch verify`
makes it clear the action being taken.

Not that `sqitch verify` exists, becuase it doesn't, yet. But I wanted to get
the name naile down now, while it's unlikely to cause anyone any pain.