Deprecated @FIRST and @LAST.
Searches for specific changes now go to the database when the context demands
it -- that is, for the `deploy` command. There is no longer a need to force
the command to search the database: it already does it.

We might want to have some way to let other commands optionally search the
database, but they also might want to search for a change as of a certain tag.
So there will have to be some other way to force a search of the database,
either through an additional option to the command or by some other means. But
there is no need for it right now. So we just deprecate these tags, but allow
them to continue to work for now.