Add `verify_change()` to Engine.
Called by `deploy_change()` if the new `with_verify` attribute is true. If the
verify fails, the deploy will be considered to have failed and the revert
process will start. If the change has no verify script a message will be sent
to STDERR but the deploy will not be considered a failure (for now). Verify
scripts will still run in log-only mode. Ref #15.