Set resolved ID for locally-resolved dependencies, too.
Was doing it only for stuff from other projects, but need to do it for local
project changes, too. Was not necessary to do this prior to b7ec1acf, because
back then, changes were deployed first. Since b7ec1acf, we check them all
ahead of time, includoing those in the current plan but not yet deployed, so
we need to populate the IDs for them, too, so that the column can be properly
set in the database.