Revert change after failed verify.
Just needed to move it down so that it runs in the same `try` block as the
logging of the change. Otherwise, when the verify failed, the change was not
rolled back, which was just bad.

While at it, customize the error message when the verify fails to say that the
verify failed, rather than leave the default error message from `Sqitch->run`.
And don't show the name of the change to be reverted in case of verify or
logging failure, as it's implicit, and was confusing when followed by a
message indicating that previous changes were being reverted.