Replace `suffix` with `rework_tags` in Plan.
When reworking a change, the code knows which tag was used for the file name
suffix, so we just assign that one tag. But when parsing a plan, we don't
know. So add all tags going back to the change being reworked, in reverse
order by change, so that we get the most recent tags first (and are more
likely have the first tag be the same as was previously used for `suffix`).