Use the current dir for the Git dir.
Because Sqitch should nearly always be called from the root directory of a
project; that's where the configuration file lives by default, for example.
May want to add an option to point to the git project directory, or we may
want to add code to search for it, depending on how things shake out.

In passing, also require that the `git` object be an instance of Git::Wrapper.