Have checkout manage multiple independent plans.
Rather than relying on the plan from the Sqitch object. This allows us to
remove the code that lets the `plan_file` attribute of Plan be either a file
handle or a file name, which felt weird. Instead, add a `parse` method to
plan, and pass the contents of the target plan to that method. Then add a
`plan` attribute to Engine (it has previously delegated the method to Sqitch)
and allow it to be read/write. The default is still to fetch it from Sqitch,
but now we can set it for other plans, as necessary. In this case, checkout
sets it before reverting, and then again after changing branches, to deploy
with the new plan.

Still need to do some error handling, and the checkout tests fail, since I did
this work on the plane and did not have Git::Wrapper installed. Will take
care of that next.