Teach `rebase` not to die if nothing needs reverting.
Instead, make a note of it and then get on with the deploy. This simplifies
the case I've often run into where I rebase, everything reverts, then the
deploy fails. I fix the failure then try to rebase again, and it would fail
because everything was already reverted. So now we just go ahead and do the
deploy. Ref issue #82.