Add `meta_destination` to engine.
And use it in the info message when the meta tables are deployed. It returns
the same value as `destination` by default, but in the SQLite engine, it
returns the stringified value of `sqitch_db`. This is because the SQLite meta
tables are not created in the same database as the changes are deployed to,
but a separate database. So it was confusing when the info message said
otherwise.