Get sqitch initialization working.
Added the live tests. Some still fail, of course, but this is a start. In
particular, all the queries that use LIMIT and/or OFFSET need to be rewritten
to use ROWNUM instead. The first query to run in the tests, in `_cid()`, has
been updated as such.

But more importantly, the loading of the sqitch schema now works. Man but
SQL*Plus is fussy! Took me forever to realize that the `/` after the CREATE
TYPE statement, which is required, would make SQL*Plus or Oracle think that it
was PL/SQL and die, because it would parse everything to that point, and there
was no COMMIT yet. So I just removed the transaction stuff. Then fixed up
all references to that type to be schema-qualified, as is the unique index
on events.

Also suprressing output in `initialize`. Nasty, though; probably should grep
through the output on error and show those lines that look like errors.