Make sure we have SQLite 3.7.11 or higher.
That was the version where the VALUES clause gained the ability to insert more
than one row. We could instead use a UNION ALL subquery, like the Oracle
engine does, and we can if anyone really complains. For now, throw an error
if DBD::SQLite is built with an older version.

While at it, fix a typo in the error message when no SQLite database is
specified.