Change sqitch_db default to $db_name\_sqitch.
So by default there will be one sqitch metadata database for every
Sqitch-managed database. I think this will be less confusing in general (I ran
into it when trying to deploy the tutorial project to multiple databases in a
single MySQL instance). As a Postgres user I think of them as more like
schemas, but for MySQL users, it think they are really more like databases, so
let's just go with it.