Go back to MySQL sqitch_db default "sqitch".
Since we need to use `checkit()` in verify tests, we need to know where it is.
That means it can't be derived from the database name (because we can't
specify a `mysql` variable with the `sqitch_db` value, as MySQL `SET`
variables are not preprocess variables for `mysql`). One can set it in the
`sqitch.conf` file to something else, of course, but then one knows what it
always is in the verify scripts.

I like it better, except for the fact that one cannot test the same Sqitch
project on a single MySQL server and use the `checkit()` function, and one
cannot deploy the same project to multiple databases on the same server
without changing `core.mysql.sqitch_db` for each one of them (and then cannot
use the `checkit()` function at all). But my hope is that deploying the same
project to multiple databases on a single MySQL server will be relatively
rare.