Add _quote_idents() to DBI Engine role.
The double-quoting of "changes" for the MySQL support broke the Oracle
support. This is because the double-quotes force case-sensitivity, and unlike
MySQL, SQLite, and Postgres, Oracle internally uses uppercase identifiers, not
lowercase. So it would have to be `"change"` for MySQL and `"CHANGE"` for
Oracle.

So instead add the `_quote_idents()` method. Pass it a list of identifiers and
it quotes any that are reserved. By default it quotes none, but the MySQL
engine will quote `changes`. Currently used only in the two places where MySQL
requires it. Its use is likely to be expanded in the future when other engines
require it such as CUBRID (issue #93).