Find the earliest change, rather than the latest.
This is to make it consistent with `$plan->find()`. It also makes sense
because, if one declares a dependency on a named change with no tag
specification, it should mean *any* instance of that change. That *does* make
it a bit harder to declare the latest instane of a change if there has been no
tag in the interrim, but one can use `@HEAD` for that. (Issue #103).