Move old templates in etc directory.
On install. This is to preserve any customizations that might have been added
in the past. The new file names will be for Postgres and SQLite; The Oracle
template is different, and likely not used by anyone much yet, so don't worry
about it. Don't bother with MySQL, either, since it hasn't been released yet.

Also requires that `process_etc_files()` recognize the old names, so that it
still appends .default. This is because the files are moved on `install`, and
`process_etc_files()` runs on `build`, so the files will not have been moved,
yet.