Teach the installer to Respect `destdir`.
Specifically when searching for existing template files. This option is mainly
used by packagers, and the truth is packages should never contain .default
files. The `destdir` directory will typically be empty when installing, so it
should be all good.

To reinforce this, the included RPM spec file now passes `destdir` to `perl
Build.PL` rather than `./Build install`. This is so that it's available
throughout the build process -- the template processing code in question is
managed by the `code` action, which comes before `install`.

While at it, don't include the `inc/Module/Build/Sqitch.pm` in the RPM (it was
included as documentation!), and tell RPM that all the etc files are
configuration files (so that RPM will place .default files as appropriate!).