Always report command errors as invalid.
Saw an example in the wild where a user inadvertently specified a command of
"foo.bar" and got the error 'Use of "require" without parentheses is
ambiguous'. She did not relize she had passed "foo.bar", and the message did
not tell her.

So change things so that the error *always* reports the passed command name,
and emit the underlying error as debugging output, so it will show up when
using `-vv`.