Avoid unexpected errors when connecting to the DB.
For Oracle and Postgres, setting the search path or current schema when the
schema does not exist should throw an error. It doesn't for me, but does
sometimes for others. In any event, I think it's a bug that it doesn't, and
have reported it to the DBI developers. So just always catch and ignore an
exception when setting the schema/search path fails. This should allow the
code that checks for a missing table error to kick in and trigger the creation
of the schema.