Move the `db_uri` attribute into Engine.
This is the only place it should ever be needed, frankly. If it's not passed
to the engine, it will be fetched from the config, and if it's not there, it
will default to "db:$engine". For back compat, it will merge values from the
engine-specific config and command-line options. One or both of those bits
will be deprecated and removed.