Rename SQLite sqitch_db attribute to "registry".
Well, it was sqitch_db_uri, but I've made "registry" a file base name. Noticed
that the code to produce the registry database name was incorrectly assuming
that the db_uri database name would always have an extension, which is not
true. So fixed that, and allowed the registry to have its own extension, in
which case it won't keep the old extension.