Move registry attribute to Engine.
Only override it in subclasses where there is some variation in the treatment
of defaults. The SQLite default does not include the deprecated config, as it
is used in `registry_uri`. That can be removed when the deprecated configs are
removed. The Oracle default is `undef`, not "sqitch", so that will be a
permanent override.

Also changed `db_uri` to use `name()` instead of `sqitch->_engine` to fetch
the name to use as the "engine" part of the URI and to fetch config values.