qHave Engine target default to config setting.
And make sure that `uri` makes use of a target value passed to the
constructor, but not falling back on its default, because the default value
builder itself falls back on `uri`, which can lead to an inifinite loop. So
use Mouse attribute accessors to only fetch the value if it is already set.
And if it's not set, still check the target config setting. This should cover
all of our bases between these two closely-related values.